refactor: move MainActivity
