+++ app/src/main/AndroidManifest.xml
@@ -16,9 +16,8 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.AtomTasks">
         <activity
-            android:name="com.costular.atomtasks.MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true"
-            android:label="@string/app_name"
             android:theme="@style/Theme.AtomTasks"
             android:windowSoftInputMode="adjustResize">
             <intent-filter>
